requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
PyMuPDF==1.23.21
Pillow==10.2.0
google-generativeai>=0.5.2
python-multipart==0.0.9
groq==0.4.2
pandas==2.2.0
python-dotenv==1.0.1
langchain-groq
reportlab==4.1.0
requests==2.31.0
langchain
langchain-core
langchain-google-genai
